+++ CHANGELOG.md
@@ -26,6 +26,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - The primary contact for the incoming trust in the export no longer uses the
   first contact if one is not set.
 - The primary contact for the incoming trust is now labelled in the export.
+- The primary contact for the outgoing trust in the export no longer uses the
+  first contact if one is not set.
+- The primary contact for the outgoing trust is now labelled in the export.
 
 ## [Release-88][release-88]